Generated Filenames Need to be Filesystem Safe
The generated filenames and folders can include characters that are
unsafe for filesystems. A list of unsafe characters can be found here,
(https://docs.microsoft.com/en-us/windows/win32/fileio/naming-a-file).
Although that specifically deals with Windows, *nix unsafe characters
are a subset of that.

When we create directory and file names, we must replace unsafe
characters with another character or remove them entirely. I believe
replacing them with a '-' is the most logical character, as it is
commonly used as a strikethrough.